Revert "(sources, overlays) Update nixpkgs{,-unstable} sources, apply perlPackages.CryptDES fix in overlay"
This reverts commit 97e4b19638b1c5a7ddc7f17c6aedfaddd8d42b4d.

Calibre-Web fails to read its library with the updated environment. I
suspect it's due the Perl-Crypt-DES update and/or my importing of the
patch, but I don't have the immediate time to dig into figuring out
which, so I'll wait and see if it occurs more generally once that patch
lands for real.